+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## CHANGELOG
 
+### v0.6.4
+
+* New json_encode command #124
+* New json_parse command #124
+* New unset_all_vars command.
+* Module level documentation #125
+* test_file command is now documented and public.
+
 ### v0.6.3 (2020-07-24)
 
 * Reduce rustc minimal version.